requirements.txt
litellm>=1.40.0
instructor>=1.3.0
mcp>=1.0.0
pydantic>=2.0
pyyaml>=6.0
jinja2>=3.1
aiofiles>=23.0
streamlit>=1.32.0
plotly>=5.18.0
python-dotenv>=1.0.0
